requirements.txt
tensorflow~=2.6.0
keras~=2.6.0
seldon-core==1.14.0
kafka-python
spicy
slack_sdk
numpy
pillow==7.2.0